Added login required decorator in sensitive functions

+++ products/views.py
@@ -1,4 +1,5 @@
 from django.shortcuts import render, get_object_or_404, redirect, reverse, HttpResponse
+from django.contrib.auth.decorators import login_required
 from django.contrib import messages
 from django.db.models.functions import Lower
 # Imports search query, q will fetch product that has search in either name or description.
@@ -66,8 +67,14 @@ def product_detail(request, product_id):
     }
     return render(request, 'products/product_detail.html', context)
 
+
+@login_required
 def add_product(request):
     """View for store owner to add products"""
+    if not request.user.is_superuser:
+        messages.error(request, 'Only store admin can do that.')
+        return redirect(reverse('home'))
+   
     if request.method == 'POST':
         form = ProductForm(request.POST, request.FILES)
         if form.is_valid():
@@ -84,10 +91,15 @@ def add_product(request):
     }
     return render(request, template, context)
 
+
+@login_required
 def edit_product(request, product_id):
     """For admin to update products"""
+    if not request.user.is_superuser:
+        messages.error(request, 'Only store admin can do that.')
+        return redirect(reverse('home'))
+
     product = get_object_or_404(Product, pk=product_id)
-    
     if request.method == 'POST':
         form = ProductForm(request.POST, request.FILES, instance=product)
         if form.is_valid():
@@ -107,8 +119,14 @@ def edit_product(request, product_id):
 
     return render(request, template, context)
 
+
+@login_required
 def delete_product(request, product_id):
     """For admin to update products"""
+    if not request.user.is_superuser:
+        messages.error(request, 'Only store admin can do that.')
+        return redirect(reverse('home'))
+        
     product = get_object_or_404(Product, pk=product_id)
     product.delete()
     messages.success(request, 'You have successfully deleted product')
